pull issues before fetch (#3)
* structure change

* pull issue

---------
