+++ estate/settings.py
@@ -33,6 +33,7 @@ ALLOWED_HOSTS = ['13.127.127.240', 'localhost',
 # Application definition
 
 INSTALLED_APPS = [
+    'agent.apps.AgentConfig',
     'property.apps.PropertyConfig',
     'django.contrib.admin',
     'django.contrib.auth',
@@ -83,30 +84,11 @@ WSGI_APPLICATION = 'estate.wsgi.application'
 # Database
 # https://docs.djangoproject.com/en/4.1/ref/settings/#databases
 
-# DATABASES = {
-#     'default': {
-#         'ENGINE': 'django.db.backends.sqlite3',
-#         'NAME': BASE_DIR / 'db.sqlite3',
-#     }
-# }
-
 DATABASES = {
     'default': {
-
-        'ENGINE': 'django.db.backends.postgresql_psycopg2',
-
-        'NAME': 'renvest_db',
-
-        'USER': 'postgres',
-
-        'PASSWORD': 'Renvest123',
-
-        'HOST': 'renvest-2023.czpy9yzm7rqx.ap-south-1.rds.amazonaws.com',
-
-        'PORT': 5432,
-
+        'ENGINE': 'django.db.backends.sqlite3',
+        'NAME': BASE_DIR / 'db.sqlite3',
     }
-
 }
