+++ property/views.py
@@ -228,148 +228,6 @@ def profile(request):
     return render(request, 'profile.html')
 
 
-def agency_registeration(request):
-    if request.user.is_authenticated:
-        if request.method == "POST":
-            agency_name = request.POST.get('agency_name')
-            agency_tagline = request.POST.get('agency_tagline')
-            agency_phone = request.POST.get('agency_phone')
-            agency_email = request.POST.get('agency_email')
-            agency_description = request.POST.get('agency_description')
-            agency_facebook = request.POST.get('agency_facebook')
-            agency_instagram = request.POST.get('agency_instagram')
-            agency_image = request.FILES['agency_image']
-
-            agency_create = Agency(user=request.user, agency_name=agency_name,
-                                   agency_tagline=agency_tagline, agency_phone=agency_phone, agency_email=agency_email,
-                                   agency_description=agency_description, agency_facebook=agency_facebook,
-                                   agency_instagram=agency_instagram, agency_image=agency_image)
-            agency_create.save()
-
-            return redirect('home')
-    else:
-        return redirect('sign-in')
-
-    return render(request, 'agency/agency-registeration.html')
-
-
-def agency_updatation(request, pk):
-    agencyDetails = 20
-    pass
-
-
-def agency_list(request):
-    agency_all = sorted(Agency.objects.all())
-    print(agency_all)
-    return render(request, 'agency/agency-list.html', {"agency_list": agency_all})
-
-
-def agency_details(request, agency_name):
-    agency_obj = Agency.objects.filter(agency_name=agency_name).first()
-    return render(request, 'agency/agency-details.html', {'agency': agency_obj})
-
-# def agency_details(request):
-#     return render(request, 'agency-details.html')
-
-
-def agent_registeration(request):
-    if request.user.is_authenticated:
-        if request.method == "POST":
-            aname = request.POST.get('name')
-            agency = request.POST.get('agency')
-            phone = request.POST.get('phone')
-            email = request.POST.get('email')
-            description = request.POST.get('description')
-            facebook = request.POST.get('facebook')
-            instagram = request.POST.get('instagram')
-            image = request.FILES['image']
-
-            agent_create = Agent(user=request.user, name=aname, agency=agency, phone=phone, email=email,
-                                 description=description, facebook=facebook, instagram=instagram, image=image)
-            agent_create.save()
-            print("Agent Create")
-            return redirect('agent-list')
-    else:
-        return redirect('sign-in')
-
-    return render(request, 'agent/agent-registeration.html')
-
-
-def agent_updatation(request, pk):
-    if request.user:
-        agentDetail = Agent.objects.filter(id=pk).first()
-        if request.method == "POST":
-            aname = request.POST.get('name')
-            agency = request.POST.get('agency')
-            phone = request.POST.get('phone')
-            email = request.POST.get('email')
-            description = request.POST.get('description')
-            facebook = request.POST.get('facebook')
-            instagram = request.POST.get('instagram')
-            image = request.FILES['image']
-
-            if image:
-                os.remove(agentDetail.image.path)
-            else:
-                image = agentDetail.image
-
-            created_at = agentDetail.created_at
-
-            agent_create = Agent(id=pk, user=request.user, name=aname, agency=agency, phone=phone, email=email,
-                                 description=description, facebook=facebook, instagram=instagram, image=image, created_at=created_at)
-            agent_create.save()
-            print("UserUpdate")
-            return redirect('agent-details', agentDetail.user.id)
-
-        elif request.user.id == agentDetail.user.id:
-            name = agentDetail.name
-            print(name)
-            agentInfo = {
-                "name": name,
-                "aname": agentDetail.agency,
-                "phone": agentDetail.phone,
-                "email": agentDetail.email,
-                "description": agentDetail.description,
-                "facebook": agentDetail.facebook,
-                "instagram": agentDetail.instagram,
-            }
-            return render(request, "agent/agent-updatation.html", agentInfo)
-
-    # print(agentInfo.values)
-    return render(request, "agent/agent-updatation.html", agentInfo)
-
-
-def agent_list(request):
-    agent_all = Agent.objects.all()
-    count = agent_all.count()
-    agent_group = []
-    for i in agent_all:
-        ap = agent_group.append(i)
-    # agent_group
-    # iterations for 6
-    variables = {
-        "count": count,
-        "agent_list": agent_group
-    }
-    return render(request, 'agent/agent-list.html', variables)
-
-
-def agent_details(request, username):
-    agent_obj = Agent.objects.filter(user=username).first()
-    if request.user.is_authenticated:
-        agentDetail = Agent.objects.filter(user=request.user).first()
-        # print(request.user.id)
-        if request.user.id == agentDetail.user.id:
-            edit = True
-    else:
-        edit = False
-    values = {
-        "edit": edit,
-        'agent': agent_obj
-    }
-    return render(request, 'agent/agent-details.html', values)
-
-
 def property_listing(request):
     if request.user.is_authenticated:
         user_agency = Agency.objects.filter(user=request.user).first()
